refactor: move Base to dedicated module
