+++ prompthelix/database.py
@@ -1,14 +1,21 @@
 import os
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
-from prompthelix.models.settings_models import APIKey
+from prompthelix.models.base import Base
 
 DATABASE_URL = os.environ.get("DATABASE_URL", "sqlite:///./prompthelix.db")
 
 engine = create_engine(DATABASE_URL)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-Base = declarative_base()
+
+__all__ = [
+    "DATABASE_URL",
+    "engine",
+    "SessionLocal",
+    "Base",
+    "get_db",
+    "init_db",
+]
 
 def get_db():
     db = SessionLocal()
