+++ prompthelix/models/__init__.py
@@ -1,4 +1,4 @@
-from prompthelix.database import Base
+from prompthelix.models.base import Base
 from .prompt_models import Prompt, PromptVersion
 
 __all__ = ["Base", "Prompt", "PromptVersion"]
